Move test trait to src/Core/test directory
