+++ tests/integration/Core/Framework/App/ActiveAppsLoaderTest.php
@@ -5,6 +5,7 @@ namespace Shopware\Tests\Integration\Core\Framework\App;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\ActiveAppsLoader;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
+use Shopware\Core\Test\AppSystemTestBehaviour;
 
 /**
  * @internal
