+++ tests/integration/Core/Framework/App/Lifecycle/AppLoaderTest.php
@@ -5,7 +5,7 @@ namespace Shopware\Tests\Integration\Core\Framework\App\Lifecycle;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\App\Manifest\Manifest;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
-use Shopware\Tests\Integration\Core\Framework\App\AppSystemTestBehaviour;
+use Shopware\Core\Test\AppSystemTestBehaviour;
 
 /**
  * @internal
